+++ dll/win32/shell32/shelldesktop/CChangeNotifyServer.cpp
