+++ dll/win32/shell32/shelldesktop/CMakeLists.txt
@@ -10,7 +10,7 @@ add_definitions(
 include_directories(${REACTOS_SOURCE_DIR}/sdk/lib/atl)
 
 list(APPEND SOURCE
-    CChangeNotify.cpp
+    CChangeNotifyServer.cpp
     CDesktopBrowser.cpp
     dde.cpp)
